requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
discord.py==1.5.1
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.10
multidict==4.7.6
playsound==1.2.2
PyAudio @ file:///PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pycparser==2.20
PyNaCl==1.4.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
urllib3==1.25.11
yarl==1.5.1
